+++ at/oneshot-roll-call.sh
@@ -1,0 +1,1 @@
+redis-cli setex stat 120 roll-call